requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
Flask==3.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
magic-filter==1.0.12
MarkupSafe==2.1.4
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
python-dateutil==2.8.2
python-dateutil==2.8.2
pytelegrambotapi
Flask==3.0.1
Flask-Login==0.6.3
arrow==1.3.0
aiogram==3.3.0
python-dotenv==1.0.1

